refactor: content-script showMessage 实现重构
