+++ src/core/bridge/background/tab.ts
@@ -7,35 +7,23 @@ export function createTabBridge(impl: ICallBridgeImpl) {
     tab: {
       async screenShot() {
         return new Promise(resolve => {
-          impl(
-            BackgroundEvents.OperateTab,
-            { type: OperateTabEnum.screenShot },
-            res => {
-              resolve(res);
-            },
-          );
+          impl(BackgroundEvents.OperateTab, { type: OperateTabEnum.screenShot }, res => {
+            resolve(res);
+          });
         });
       },
       async getCurrent(): Promise<chrome.tabs.Tab | undefined> {
         return new Promise(resolve => {
-          impl(
-            BackgroundEvents.OperateTab,
-            { type: OperateTabEnum.getCurrent },
-            res => {
-              resolve(res);
-            },
-          );
+          impl(BackgroundEvents.OperateTab, { type: OperateTabEnum.getCurrent }, res => {
+            resolve(res);
+          });
         });
       },
       async create(url: string) {
         return new Promise(resolve => {
-          impl(
-            BackgroundEvents.OperateTab,
-            { type: OperateTabEnum.create, url },
-            res => {
-              resolve(res);
-            },
-          );
+          impl(BackgroundEvents.OperateTab, { type: OperateTabEnum.create, url }, res => {
+            resolve(res);
+          });
         });
       },
       async getDocument(): Promise<{
@@ -44,13 +32,23 @@ export function createTabBridge(impl: ICallBridgeImpl) {
         title: string;
       }> {
         return new Promise(resolve => {
-          impl(
-            BackgroundEvents.OperateTab,
-            { type: OperateTabEnum.getDocument },
-            res => {
-              resolve(res);
-            },
-          );
+          impl(BackgroundEvents.OperateTab, { type: OperateTabEnum.getDocument }, res => {
+            resolve(res);
+          });
+        });
+      },
+      async showMessage(data: {
+        type: 'error' | 'success';
+        text: string;
+        link?: {
+          text: string;
+          href: string;
+        };
+      }) {
+        return new Promise(resolve => {
+          impl(BackgroundEvents.OperateTab, { type: OperateTabEnum.showMessage, data }, res => {
+            resolve(res);
+          });
         });
       },
     },
