+++ src/isomorphic/background/tab.ts
@@ -3,9 +3,11 @@ export enum OperateTabEnum {
   getCurrent = 'getCurrent',
   create = 'create',
   getDocument = 'getDocument',
+  showMessage = 'showMessage',
 }
 
 export interface IOperateTabData {
   type: OperateTabEnum;
   url?: string;
+  data?: any;
 }
