+++ src/isomorphic/event/contentScript.ts
@@ -3,20 +3,4 @@ export enum ContentScriptEvents {
   LevitateConfigChange = 'contentScript/levitateConfigChange',
   ForceUpgradeVersion = 'contentScript/forceUpgradeVersion',
   LoginOut = 'contentScript/LoginOut',
-  GetDocument = 'contentScript/getDocument',
-}
-
-export const ContentScriptMessageKey = 'ContentScriptMessageKey';
-
-export enum ContentScriptMessageActions {
-  ShowMessage = 'showMessage',
-}
-
-export interface IShowMessageData {
-  type: 'error' | 'success';
-  text: string;
-  link?: {
-    text: string;
-    href: string;
-  };
 }
