+++ src/pages/inject/WordMark/index.module.less
@@ -29,4 +29,3 @@
     outline-style: none;
   }
 }
-
