+++ src/pages/inject/action-listener.tsx
@@ -1,11 +1,4 @@
-import React from 'react';
-import { message } from 'antd';
-import {
-  ContentScriptEvents,
-  ContentScriptMessageKey,
-  ContentScriptMessageActions,
-  IShowMessageData,
-} from '@/isomorphic/event/contentScript';
+import { ContentScriptEvents } from '@/isomorphic/event/contentScript';
 import { WordMarkMessageActions } from '@/isomorphic/event/wordMark';
 import { AccountLayoutMessageActions } from '@/isomorphic/event/accountLayout';
 import { App } from './content-scripts';
@@ -22,51 +15,28 @@ export interface RequestMessage<T> {
 
 export const initContentScriptActionListener = (context: App) => {
   chrome.runtime.onMessage.addListener(
-    (
-      request: RequestMessage<any>,
-      _sender: MessageSender,
-      sendResponse: SendResponse,
-    ) => {
+    (request: RequestMessage<any>, _sender: MessageSender, sendResponse: SendResponse) => {
       switch (request.action) {
         case ContentScriptEvents.WordMarkConfigChange: {
-          context.sendMessageToWordMark(
-            WordMarkMessageActions.wordMarkConfigUpdate,
-            request.data,
-          );
+          context.sendMessageToWordMark(WordMarkMessageActions.wordMarkConfigUpdate, request.data);
           sendResponse(true);
           break;
         }
         case ContentScriptEvents.LevitateConfigChange: {
-          context.sendMessageToLevitateBall(
-            LevitateBallMessageActions.levitateBallConfigUpdate,
-            request.data,
-          );
+          context.sendMessageToLevitateBall(LevitateBallMessageActions.levitateBallConfigUpdate, request.data);
           sendResponse(true);
           break;
         }
         case ContentScriptEvents.ForceUpgradeVersion: {
-          context.sendMessageToAccountLayout(
-            AccountLayoutMessageActions.ForceUpdate,
-            {
-              html: request.data?.html,
-            },
-          );
+          context.sendMessageToAccountLayout(AccountLayoutMessageActions.ForceUpdate, {
+            html: request.data?.html,
+          });
           sendResponse(true);
           break;
         }
         case ContentScriptEvents.LoginOut: {
-          context.sendMessageToAccountLayout(
-            AccountLayoutMessageActions.LoginOut,
-          );
+          context.sendMessageToAccountLayout(AccountLayoutMessageActions.LoginOut);
           sendResponse(true);
-          break;
-        }
-        case ContentScriptEvents.GetDocument: {
-          sendResponse({
-            url: window.location.href,
-            html: document.documentElement.outerHTML,
-            title: document.title,
-          });
           break;
         }
         default: {
@@ -76,48 +46,4 @@ export const initContentScriptActionListener = (context: App) => {
       return true;
     },
   );
-};
-
-export const initContentScriptMessageListener = () => {
-  window.addEventListener('message', (e: MessageEvent<any>) => {
-    const { action, key, data } = e.data || {};
-    if (key !== ContentScriptMessageKey) {
-      return;
-    }
-    switch (action) {
-      case ContentScriptMessageActions.ShowMessage: {
-        const { type, text, link } = (data || {}) as IShowMessageData;
-        const content = (
-          <span className="yuque-chrome-extension-message">
-            <span className="yuque-chrome-extension-message-text">{text}</span>
-            {!!link && (
-              <a
-                target="_blank"
-                href={link.href}
-                className="yuque-chrome-extension-message-href"
-              >
-                {link.text}
-              </a>
-            )}
-          </span>
-        );
-        if (type === 'success') {
-          message.success({
-            content,
-            className: 'yuque-chrome-extension-message-wrapper',
-          });
-        }
-        if (type === 'error') {
-          message.error({
-            content,
-            className: 'yuque-chrome-extension-message-wrapper',
-          });
-        }
-        break;
-      }
-      default: {
-        break;
-      }
-    }
-  });
 };
