--- src/styles/inject.less
@@ -1,27 +1,0 @@
-@import '~@/styles/parameters.less';
-
-.yuque-chrome-extension-message-wrapper {
-  .ant-message-custom-content {
-    display: flex;
-
-    .anticon-check-circle {
-      margin-top: 2px;
-    }
-
-    .yuque-chrome-extension-message {
-      display: flex;
-      font-size: @font-size;
-    }
-
-    .yuque-chrome-extension-message-text {
-      font-size: @font-size;
-    }
-
-    .yuque-chrome-extension-message-href {
-      color: @link-color;
-      outline: none;
-      text-decoration: none;
-      margin-left: 8px;
-    }
-  }
-}
